move tag validation to the type crate
